Rename design rules and tests for consistency with rule names
