+++ src/Rule/Design/ExcessiveMethodLength.php
@@ -27,7 +27,7 @@ use PHPMD\Rule\MethodAware;
  * This rule will detect to long methods, those methods are unreadable and in
  * many cases the result of copy and paste coding.
  */
-final class LongMethod extends AbstractRule implements FunctionAware, MethodAware
+final class ExcessiveMethodLength extends AbstractRule implements FunctionAware, MethodAware
 {
     /**
      * This method checks the lines of code length for the given function or
