+++ tests/php/PHPMD/Rule/Design/ExcessiveMethodLengthTest.php
@@ -24,8 +24,8 @@ use PHPUnit\Framework\Attributes\CoversClass;
 /**
  * Test case for the excessive long method rule.
  */
-#[CoversClass(LongMethod::class)]
-class LongMethodTest extends AbstractTestCase
+#[CoversClass(ExcessiveMethodLength::class)]
+class ExcessiveMethodLengthTest extends AbstractTestCase
 {
     /**
      * Tests that the rule applies for a value greater than the configured
@@ -36,7 +36,7 @@ class LongMethodTest extends AbstractTestCase
         $method = $this->getMethodMock('loc', 42);
         $report = $this->getReportWithOneViolation();
 
-        $rule = new LongMethod();
+        $rule = new ExcessiveMethodLength();
         $rule->setReport($report);
         $rule->addProperty('minimum', '41');
         $rule->addProperty('ignore-whitespace', '0');
@@ -52,7 +52,7 @@ class LongMethodTest extends AbstractTestCase
         $method = $this->getMethodMock('loc', 42);
         $report = $this->getReportWithOneViolation();
 
-        $rule = new LongMethod();
+        $rule = new ExcessiveMethodLength();
         $rule->setReport($report);
         $rule->addProperty('minimum', '42');
         $rule->addProperty('ignore-whitespace', '0');
@@ -68,7 +68,7 @@ class LongMethodTest extends AbstractTestCase
         $method = $this->getMethodMock('loc', 22);
         $report = $this->getReportWithNoViolation();
 
-        $rule = new LongMethod();
+        $rule = new ExcessiveMethodLength();
         $rule->setReport($report);
         $rule->addProperty('minimum', '23');
         $rule->addProperty('ignore-whitespace', '0');
@@ -83,7 +83,7 @@ class LongMethodTest extends AbstractTestCase
         $class = $this->getClassMock('eloc', 22);
         $report = $this->getReportWithNoViolation();
 
-        $rule = new LongMethod();
+        $rule = new ExcessiveMethodLength();
         $rule->setReport($report);
         $rule->addProperty('minimum', '23');
         $rule->addProperty('ignore-whitespace', '1');
